move around model code
